+++ hooman/__init__.py
@@ -1,5 +1,5 @@
 from .hooman import Hooman
 from .formula import *
 
-version_info = (0, 8, 0)
+version_info = (0, 8, 2)
 __version__ = ".".join([str(v) for v in version_info])